botim_django/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-parler==2.3
djangorestframework==3.14.0
pytz==2023.3.post1
shortuuid==1.0.11
sqlparse==0.4.4
